Major refactor: - username self-defined and email updatable - id int instead of ulong

+++ Application/LogicInterfaces/IUserLogic.cs
@@ -13,6 +13,6 @@ namespace Application.LogicInterfaces
         Task<User> CreateAsync(UserCreationDTO dto);
         Task<User> LoginAsync(UserLoginDTO dto);
         Task UpdateAsync(UserUpdateDTO dto);
-        Task DeleteAsync(ulong id);
+        Task DeleteAsync(int id);
     }
 }
